backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.1

# Database
psycopg2-binary==2.9.11
sqlalchemy==2.0.36
alembic==1.14.0

# Vector Database
pgvector==0.3.6

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20

# AI/ML
langchain==0.3.13
langchain-openai==0.2.14
langchain-community==0.3.13
openai==1.59.6

# Document Processing
pypdf==5.1.0
python-docx==1.1.2

# Utilities
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.7.0
email-validator==2.2.0

# Rate Limiting
slowapi==0.1.9

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1

# Additional utilities
aiofiles==24.1.0